Changes based on a suggestion by kevinlekiller "Instead of requerying, the query on line 249 can be changed to select the categoryid."
This meant changing it to an in clause as well, as such I also added technical books to the list since lookup of them on amazon should also work if published.

This also mean that groups such a foreign books could be handled as well, but I do not think this should be done without UI changes which I did not want to make in this commit.  In theory this could be changes in teh future to allow a selection of categories from the UI as to what books to process.