Added a number of new groups:         alt.binaries.multimedia.erotica.anime         alt.binaries.multimedia.erotica.asian         alt.binaries.town.xxx         alt.binaries.mp3.audiobooks.scifi-fantasy         alt.binaries.sounds.audiobooks.scifi-fantasy         alt.binaries.mp3.abooks         alt.binaries.nzbpirates         alt.binaries.town.cine         alt.binaries.usenet-space-cowboys         alt.binaries.warez.ibm-pc.games         alt.binaries.warez.games         alt.binaries.software         alt.binaries.e-book.magazines         alt.binaries.sounds.anime
Reviewing usenet groups between 1/12 and /17 and looking at groups that had a number of post which were not present.
Also looking for groups with valid posts (nzbpirates is a very mixed bag), but the others contained many-many valid posts.
I also added 2 slow audiobook groups and a new magazine ebook group (which while small has a very number of new posts)

Finally I added nameCleaning functions and if-else statements for the new groups.  This is almost all cut and paste from existing groups, since the content is similar, however in this way the required statements are present and have a good starting point for modifications to be made.

The PGSQL script is designed based upon best practices for pgsql and duplicate inserts, but as I do not have this DB I was unable to properly test this, I noticed in the past the new group patches were missing for pgsql, so there was no basis to work from.