Split NameCleaning into 2 files CollectionsCleaning and ReleaseCleaning No changes were made to either file I simple split them into 2.
I did make a few comments on unused code, or code which appears to be unknown, but I did not change any code as I'm only refactoring.

Overall the conversion was reasonable easy and straight forward, all comparisions are straight forward, except for the CollectionsCleaning and ReleaseCleaning, however a simple diff shows that there are very no code changes, only the fact it has been split into 2 files and the function are restricted to those called.