Refectored ReleaseCleaning to match the case structure of Collection cleaning, to make it easier to work with. While I was at it I also ensured that all groups are covered in each file (if there is no entry then they use generic), again this should make it easier to maintain, and dsince they are set to generic which is what they would have matched there is no change to the call tree. I also set the function names in ReleaseCleaning to match those from Collections cleaning, again this should help make life easier.
I changed no functional code or the logic, only converted a from an if/else to a case statement
This does mean that ReleaseCleaning.php is very different looking now, since I also needed to change the name $subject to $this->subject, but I tried to leave the spacing the same, so it is easy to see that nothing regex or functionality related changed.

I'm sorry this required such a large change, but then again there is no way to make is simple as the entire if/else needed to be converted to functions.