Proposed TPG SP
terbonium - see what you think about this.  The SP Works, not sure if it will work with our autopatcher yet (that's why it is on one line).  If the user isn't TPG it won't find any tables LIKE binaries_% and should not run.  Additionally, we could modify it so that it does binaries as well as the TPG ones in one loop.