Fix `zed update nzedb`
Updating was failing prematurely because `Update::nzedb()` expects `Update::git()` to return the output of the `Update::git->pull()`. This commit adjusts `Update::git()` to not only output the response from `Update::git->pull()`, but also to return the trimmed value for evaluation in `Update::nzedb()`, allowing the update to continue.