Initial commit of web page refactor.
It quickly became obvious refactoring old smarty pages to use Cake's
models would take as much work as rewriting the pages completely,
possibly more.

The theme is based on AdminLTE (has a plugin for it as a dependency),
and should allow me to get a basic but functional set of pages working.

Will start on admin pages next.